+++ inst/ui-editor-react/src/Shiny-Ui-Elements/commonSettingsTemplates.tsx
@@ -1,11 +1,5 @@
 import type { DynamicFieldInfoByType } from "components/Inputs/SettingsFormBuilder/inputFieldTypes";
 
-export const outputIdInfo: DynamicFieldInfoByType["string"] = {
-  label: "Output ID",
-  inputType: "string",
-  defaultValue: "outputId",
-};
-
 function makeStringInputInfo(
   label: string,
   defaultValue: string
@@ -15,30 +9,6 @@ function makeStringInputInfo(
     inputType: "string",
     defaultValue: defaultValue,
   };
-}
-
-export function stringInput(
-  args: Omit<DynamicFieldInfoByType["string"], "inputType">
-): DynamicFieldInfoByType["string"] {
-  return {
-    inputType: "string",
-    ...args,
-  };
-}
-
-/**
- * Make info object for the common `outputId` argument
- * @param defaultValue Default value for `outputId` input
- * @returns String input info object with label of "Output ID" and provided
- * default value
- */
-export function makeOutputIdInfo(
-  defaultValue: string
-): DynamicFieldInfoByType["string"] {
-  return stringInput({
-    label: "Output ID",
-    defaultValue,
-  });
 }
 
 /**
@@ -64,29 +34,3 @@ export function makeLabelInputInfo(
 ): DynamicFieldInfoByType["string"] {
   return makeStringInputInfo("Label text", defaultValue);
 }
-
-export const inputIdInfo: DynamicFieldInfoByType["string"] = {
-  label: "Input ID",
-  inputType: "string",
-  defaultValue: "inputId",
-};
-
-function makeWidthInfo({
-  optional,
-}: {
-  optional: boolean;
-}): DynamicFieldInfoByType["cssMeasure"] {
-  return {
-    inputType: "cssMeasure",
-    label: "Width",
-    defaultValue: "100%",
-    units: ["%", "px", "rem"],
-    ...(optional ? { optional: true } : {}),
-  };
-}
-
-export const requiredWidthInfo: DynamicFieldInfoByType["cssMeasure"] =
-  makeWidthInfo({ optional: false });
-
-export const optionalWidthInfo: DynamicFieldInfoByType["cssMeasure"] =
-  makeWidthInfo({ optional: true });